Fix libusb context abuses causing segfaults on OpenWrt
See http://forum.kryoflux.com/viewtopic.php?f=3&t=358 for more
information.

This is the _lazy_ way of fixing it, by making the manufacturing library
only capable of a single libusb context.  The _good_ way would be to
make all uses of libusb internal to this library use the context
properly.